fileconvertbot/requirements.txt
python-telegram-bot==20.3
Pillow==10.0.0
img2pdf==0.4.4
python-docx==0.8.11
PyPDF2==3.0.1
reportlab==4.0.4